major spatial rework

+++ crates/feagi-basis/feagi-basis-collections/src/spatial_indexing_structs/axis_order.rs
@@ -7,12 +7,12 @@ pub type AxisOrderIdentifier = u8;
 #[derive(Debug, PartialEq, Eq, Hash, Clone, Copy, ::serde::Serialize, ::serde::Deserialize)]
 pub struct AxisOrderArray<const NUM_DIMS: usize>{
     #[serde(with = "serde_arrays")]
-    data: [usize; NUM_DIMS]
+    data: [u8; NUM_DIMS]
 }
 
 impl<const NUM_DIMS: usize> AxisOrderArray<NUM_DIMS> {
     /// Borrow dimensions as a fixed-size slice.
-    pub fn as_slice(&self) -> &[usize; NUM_DIMS] {
+    pub fn as_slice(&self) -> &[u8; NUM_DIMS] {
         &self.data
     }
 }
@@ -30,31 +30,31 @@ pub enum AxisOrderEnum {
 impl AxisOrderEnum {
     
     pub const fn generate_axis_order_array<const NUM_DIMS: usize>(self) -> AxisOrderArray<NUM_DIMS> {
-        let mut out = [0usize; NUM_DIMS];
-        let mut i = 0usize;
+        let mut out = [0u8; NUM_DIMS];
+        let mut i = 0u8;
         match self { 
             AxisOrderEnum::DefaultIncrementing => {
-                while i < NUM_DIMS {
-                    out[i] = i;
+                while i < NUM_DIMS as u8 {
+                    out[i as usize] = i;
                     i = i + 1;
                 }
             }
             AxisOrderEnum::Decrementing => {
-                i = NUM_DIMS;
+                i = NUM_DIMS as u8;
                 while i != 0 {
                     i = i - 1;
-                    out[i] = i
+                    out[i as usize] = i
                 }
             }
         }
         AxisOrderArray {data: out}
     }
 
-    pub const fn from_identifier<const NUM_DIMS: usize>(u: AxisOrderIdentifier)  -> AxisOrderEnum {
+    pub const fn from_identifier<const NUM_DIMS: u8>(u: AxisOrderIdentifier)  -> AxisOrderEnum {
         Self::from_u8::<NUM_DIMS>(u)
     }
 
-    pub const fn from_u8<const NUM_DIMS: usize>(u: u8) -> AxisOrderEnum {
+    pub const fn from_u8<const NUM_DIMS: u8>(u: u8) -> AxisOrderEnum {
         match u { 
             0 => AxisOrderEnum::DefaultIncrementing,
             1 => AxisOrderEnum::Decrementing,
